+++ z/util/regex/rules/orlist.hpp
@@ -20,7 +20,7 @@ namespace z
 				 * \param max The maximum number of times this rule can match.
 				 * \param greedy Whether this rule should consume input greedily.
 				 */
-				orlist(bool negate=false, size_t min=1, size_t max=1, bool greedy=true);
+				orlist(bool negate=false, int min=1, int max=1, bool greedy=true);
 				bool match(core::inputStream& stream) const;
 				bool isOr() const {return true;}
